+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed RFC3339 timestamp validation for policy expirations and agent protocol
+  deadlines.
+
 ## [0.7.1] - 2026-04-28
 
 ### Fixed